mdq-spec/pyproject.toml
[project]

name = "mdq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fábio Macêdo Mendes", email = "fabiomacedomendes@gmail.com" },
]
requires-python = ">=3.13"
dependencies = [
    "datamodel-code-generator>=0.30.0",
    "invoke>=2.2.0",
    "jsonref>=1.1.0",
    "jsonschema>=4.23.0",
    "rich>=14.0.0",
    "ruamel-yaml>=0.18.10",
    "tyro>=0.9.18",
]

[project.scripts]
mdq-transform = "mdq_schema.cli.transforms:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
